refactor(js): cleanup RN side
