+++ wrappers/javascript/react-native/src/ReactNativeIndyVdr.ts
