+++ wrappers/javascript/react-native/src/index.ts
@@ -1,11 +1,12 @@
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+
 import { NativeModules } from 'react-native'
 
-import { ReactNativeIndyVdr } from './indyVdr'
+import { ReactNativeIndyVdr } from './ReactNativeIndyVdr'
 
 const module = NativeModules.IndyVdr
-const res = module.install()
-if (!res) {
-  throw Error('Unable to install the turboModule: indyVdr')
-}
+if (!module.install()) throw Error('Unable to install the turboModule: indyVdr')
 
 export const indyVdrReactNative = new ReactNativeIndyVdr()
