+++ wrappers/javascript/react-native/src/library/index.ts
@@ -1,0 +1,1 @@
+export * from './register'
