+++ wrappers/javascript/react-native/src/library/register.ts
